pyproject.toml
[project]
name = "origami-wsi-tools"
version = "0.1.0"
description = "WSI residual pyramid CLI tools for ORIGAMI."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyvips",
  "Pillow",
  "numpy",
  "scikit-image",
  "openslide-python",
  "requests",
  "tqdm",
  "ipykernel>=7.1.0",
  "scipy",
  "matplotlib",
  "torch",
  "lpips",
  "sewar",
  "lz4",
  "opencv-contrib-python>=4.13.0.92",
  "pywavelets>=1.8.0",
]

[project.scripts]
wsi-residual-tool = "cli.wsi_residual_tool:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["cli"]

[tool.uv]
package = true
